Implement comment submission on reports
I don't update the 'user' collection with this comment yet; I'm planning on discussing with my group members whether tracking comments in the 'user' collection is necessary.

Also probably on the to-do list is updating the comment list client-side instead of forcing the page to reload using a redirect.

Also also, this commit includes some error class cleanups.